Varswap more-less done.

+++ src/lib/diagrams/bdd_manager_tools.ipp
@@ -18,7 +18,7 @@ namespace mix::dd
         (bdd_t& d) -> double
     {
         return static_cast<double>(this->satisfy_count(d))
-             / utils::two_pow(this->get_var_count());
+             / static_cast<double>(utils::two_pow(this->get_var_count()));
     }
 
     template<class VertexData, class ArcData>
